Test utility classes are moved to util package
