+++ src/test/java/gov/ca/cwds/idm/service/IdmServiceImplTest.java
@@ -4,8 +4,8 @@ import static gov.ca.cwds.BaseIntegrationTest.H2_DRIVER_CLASS_NAME;
 import static gov.ca.cwds.BaseIntegrationTest.IDM_BASIC_AUTH_PASS;
 import static gov.ca.cwds.BaseIntegrationTest.IDM_BASIC_AUTH_USER;
 import static gov.ca.cwds.idm.service.IdmServiceImpl.transformSearchValues;
-import static gov.ca.cwds.idm.service.TestHelper.user;
-import static gov.ca.cwds.idm.service.TestHelper.userType;
+import static gov.ca.cwds.idm.util.TestHelper.user;
+import static gov.ca.cwds.idm.util.TestHelper.userType;
 import static gov.ca.cwds.idm.service.cognito.StandardUserAttribute.EMAIL;
 import static gov.ca.cwds.idm.service.cognito.StandardUserAttribute.FIRST_NAME;
 import static gov.ca.cwds.idm.service.cognito.StandardUserAttribute.RACFID_STANDARD;
@@ -26,7 +26,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
 import com.amazonaws.services.cognitoidp.model.UserType;
-import gov.ca.cwds.idm.WithMockCustomUser;
+import gov.ca.cwds.idm.util.WithMockCustomUser;
 import gov.ca.cwds.idm.dto.User;
 import gov.ca.cwds.idm.dto.UserEnableStatusRequest;
 import gov.ca.cwds.idm.dto.UserUpdate;
