+++ src/test/java/gov/ca/cwds/idm/service/authorization/AuthorizationServiceImplTest.java
@@ -4,8 +4,8 @@ import static gov.ca.cwds.config.api.idm.Roles.COUNTY_ADMIN;
 import static gov.ca.cwds.config.api.idm.Roles.CWS_WORKER;
 import static gov.ca.cwds.config.api.idm.Roles.OFFICE_ADMIN;
 import static gov.ca.cwds.config.api.idm.Roles.STATE_ADMIN;
-import static gov.ca.cwds.idm.service.TestHelper.admin;
-import static gov.ca.cwds.idm.service.TestHelper.user;
+import static gov.ca.cwds.idm.util.TestHelper.admin;
+import static gov.ca.cwds.idm.util.TestHelper.user;
 import static gov.ca.cwds.util.CurrentAuthenticatedUserUtil.getCurrentUser;
 import static gov.ca.cwds.util.CurrentAuthenticatedUserUtil.getCurrentUserCountyName;
 import static gov.ca.cwds.util.CurrentAuthenticatedUserUtil.getCurrentUserOfficeIds;
